Pull RDMA git-fixes from Nicolas Morey

suse-commit: 73650cf262f372d25401db0085ae9d573f6465c9

+++ drivers/infiniband/hw/irdma/verbs.h
@@ -117,7 +117,7 @@ struct irdma_cq {
 	struct irdma_sc_cq sc_cq;
 	u16 cq_head;
 	u16 cq_size;
-	u16 cq_num;
+	u32 cq_num;
 	bool user_mode;
 	atomic_t armed;
 	enum irdma_cmpl_notify last_notify;
